1.Update the test set. 2.Add freesurfer affine evaluation script. 3.Move all the helper functions to helper.py. 4.Add a prepare script so that we process the image for Synthmorph once.
